Dish to product

+++ app/src/main/java/model/Skladnik.kt
@@ -3,10 +3,10 @@ package com.example.softwareengineering.model
 data class Skladnik(
     val id: String? = "",
     val name: String = "",
-    val calories: Int = 0,
-    val protein: Int = 0,
-    val carbs: Int = 0,
-    val fat: Int = 0,
+    val calories: Double = 0.0,
+    val protein: Double = 0.0,
+    val carbs: Double = 0.0,
+    val fat: Double = 0.0,
     val userId: String? = "",
     var checked: Boolean = false
 )
